Improve: Badge 컴포넌트 개선
