+++ Umpa/Presentation/UIHelpers/CustomModifiers.swift
@@ -1,0 +1,12 @@
+// Created for Umpa in 2025
+
+import SwiftUI
+
+extension View {
+    @inlinable func innerStroke<S>(_ shapeStyle: S, cornerRadius: CGFloat) -> some View where S: ShapeStyle {
+        self.overlay(
+            RoundedRectangle(cornerRadius: cornerRadius)
+                .strokeBorder(shapeStyle, lineWidth: 1)
+        )
+    }
+}
